requirements.txt
requests>=2.25.1
python-dotenv>=0.19.0
pytest>=6.0.0
flask>=2.0.0
flask-cors>=3.0.10
jinja2>=3.0.0
